# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "base model for any failed / error API response",
        "title": "ATEC API Docs",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/auth/login": {
            "post": {
                "description": "Use this endpoint to login with your username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Gain access to the system by authenticating using a registered account",
                "parameters": [
                    {
                        "description": "account detail such as email and password to log in",
                        "name": "login_input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.LoginOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication Failed",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/password": {
            "get": {
                "description": "After user click the link from email for reset password, he will be redirected here to change the password. This might not work in this page\nsince the returned value in this API is HTML",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Webpage's form to change user's account password",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request\"\t\"validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "After using init reset password method, the resulting JWT token can be used here\nto authorize the password changes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Change user's account password",
                "parameters": [
                    {
                        "description": "the email of the account which password will be reset",
                        "name": "change_password_input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.ResetPasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.ResetPasswordOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request\"\t\"validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "If the user wants to change their password, use this API.\nwhen the request succeed, an email containing confirmation link will\nbe sent to the account email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Initiate change password process for an active account",
                "parameters": [
                    {
                        "description": "the email of the account which password will be reset",
                        "name": "init_change_password_input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.InitResetPasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.InitResetPasswordOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request\"\t\"validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/signup": {
            "post": {
                "description": "Use this API endpoint to create a new account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a new account to access user only resources within this system",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "signup_input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.SignupInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.SignupOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request / validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/signup/resend": {
            "post": {
                "description": "If something happen during account verification that cause the email not received\nor the email is lost, use this API to resend the verification email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Resend email for account verification",
                "parameters": [
                    {
                        "description": "resend signup input",
                        "name": "resend_signup_input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.ResendVerificationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.ResendVerificationOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request / validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rest.InitResetPasswordInput": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "rest.InitResetPasswordOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "rest.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "rest.LoginOutput": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "rest.ResetPasswordInput": {
            "type": "object",
            "required": [
                "new_password",
                "token"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "rest.ResetPasswordOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rest.SignupOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "confirmation link sent to your email"
                }
            }
        },
        "rest.StandardErrorResponse": {
            "description": "base model for any failed / error API response",
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string",
                    "example": "missing required fields on input"
                },
                "error_message": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "rest.StandardSuccessResponse": {
            "description": "base model for any successful API response",
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminLevelAuth": {
            "description": "Bearer Token authentication for secure endpoints accessible only by Admin level user",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserLevelAuth": {
            "description": "Bearer Token authentication for secure endpoints accessible only by registered user with auth token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}